pyproject.toml
[tool.poetry]
name = "maa-api"
version = "0.1.0"
description = ""
authors = ["Chenzhipeng <2010768120@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
jinja2 = "^3.0.0"
adbutils = "^2.8.0"
ruamel-yaml = "^0.18.7"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
